third_party/WebKit/LayoutTests/html/text_level_semantics/a-click.html

<!DOCTYPE HTML>
<body/>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<a id="target" href="#abc" title="This should appear as a tooltip">Hello.</a>
<script>
function doubleClickNavigation() {
  document.getElementById('target').dispatchEvent(new MouseEvent("click", {detail: 2}));
  assert_equals(window.location.hash, "#abc", "Click will navigate on double click");
}

test(doubleClickNavigation, "Double Click Navigation");
</script>
